# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Koupreng E-Invitation API
  description: |
    REST API for the Koupreng E-Invitation platform, including authentication,
    users, invitations, templates, guests, RSVP, check-in, media, payments,
    organizations, subscriptions, notifications, reporting, administration,
    and related services. The current contract includes both legacy /api routes
    and versioned /api/v1 routes; the version below does not imply a global base path.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/invitations/{invitationId}/guests/{guestId}/check-in:
    post:
      tags:
      - Check-In
      summary: Check in a guest manually
      operationId: manual
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: guestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualCheckInRequest"
            examples:
              manualCheckIn:
                summary: Manually check in a guest
                value:
                  note: Verified by the reception team
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCheckInResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Check-In
      summary: Undo
      operationId: undo
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: guestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/check-in/scan:
    post:
      tags:
      - Check-In
      summary: Process a guest check-in scan
      operationId: scan
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInScanRequest"
            examples:
              fixtureGuest:
                summary: Check in the attending fixture guest
                value:
                  token: demo-guest-attending-token
                  note: Checked in from the Scalar API reference
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCheckInResponse"
      security:
      - bearerAuth: []
...
  /api/v1/notifications/summary:
    get:
      tags:
      - Notifications
      summary: Summarize notification activity
      operationId: summary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseNotificationSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/seating/summary:
    get:
      tags:
      - Seating
      summary: Summarize seating activity
      operationId: summary_1
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeatingSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/rsvps/summary:
    get:
      tags:
      - RSVP
      summary: Summarize rsvp activity
      operationId: summary_2
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/check-in/summary:
    get:
      tags:
      - Check-In
      summary: Summarize check in activity
      operationId: summary_4
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCheckInSummaryResponse"
      security:
      - bearerAuth: []
  /api/v1/invitations/{invitationId}/check-in/list:
    get:
      tags:
      - Check-In
      summary: List check in records
      operationId: list_5
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListCheckInResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/analytics/check-in:
    get:
      tags:
      - Administration
      summary: Analytics Check In
      operationId: analyticsCheckIn
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminReportResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ManualCheckInRequest:
      type: object
      properties:
        note:
          type: string
          maxLength: 1000
          minLength: 0
    ApiResponseCheckInResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/CheckInResponse"
        timestamp:
          type: string
          format: date-time
    CheckInResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        invitationId:
          type: integer
          format: int64
        guestId:
          type: integer
          format: int64
        guestName:
          type: string
        checkedInAt:
          type: string
          format: date-time
        checkedInByUserId:
          type: integer
          format: int64
        source:
          type: string
        note:
          type: string
        alreadyCheckedIn:
          type: boolean
        result:
          type: string
...
    CheckInScanRequest:
      type: object
      properties:
        token:
          type: string
          maxLength: 1200
          minLength: 0
        note:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - token
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ApiResponseRsvpSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/RsvpSummaryResponse"
        timestamp:
          type: string
          format: date-time
    RsvpSummaryResponse:
      type: object
      properties:
        totalGuests:
          type: integer
          format: int64
        attending:
          type: integer
          format: int64
        notAttending:
          type: integer
          format: int64
        maybe:
          type: integer
          format: int64
        pending:
          type: integer
          format: int64
        totalAttendeeCount:
          type: integer
          format: int64
...
    ApiResponseNotificationSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/NotificationSummaryResponse"
        timestamp:
          type: string
          format: date-time
    NotificationSummaryResponse:
      type: object
      properties:
        total:
          type: integer
          format: int64
        unread:
          type: integer
          format: int64
        pending:
          type: integer
          format: int64
        sent:
          type: integer
          format: int64
        delivered:
          type: integer
          format: int64
        failed:
          type: integer
          format: int64
        cancelled:
          type: integer
          format: int64
...
    ApiResponseSeatingSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/SeatingSummaryResponse"
        timestamp:
          type: string
          format: date-time
    SeatingSummaryResponse:
      type: object
      properties:
        invitationId:
          type: integer
          format: int64
        totalTables:
          type: integer
          format: int64
        totalCapacity:
          type: integer
          format: int64
        assignedSeats:
          type: integer
          format: int64
        remainingSeats:
          type: integer
          format: int64
...
    ApiResponseCheckInSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/CheckInSummaryResponse"
        timestamp:
          type: string
          format: date-time
    CheckInSummaryResponse:
      type: object
      properties:
        invitationId:
          type: integer
          format: int64
        totalGuests:
          type: integer
          format: int64
        checkedIn:
          type: integer
          format: int64
        remaining:
          type: integer
          format: int64
        attendingCheckedIn:
          type: integer
          format: int64
    ApiResponseListCheckInResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/CheckInResponse"
        timestamp:
          type: string
          format: date-time
...
    AdminReportResponse:
      type: object
      properties:
        report:
          type: string
        generatedAt:
          type: string
          format: date-time
        summary:
          type: object
          additionalProperties: {}
        rows:
          type: array
          items: {}
    ApiResponseAdminReportResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/AdminReportResponse"
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste the accessToken returned by POST /api/auth/login. Scalar
        adds the Bearer prefix automatically; do not type it twice.
      scheme: bearer
      bearerFormat: JWT
    adminPaymentSecret:
      type: apiKey
      description: Internal Telegram payment reconciliation secret.
      name: X-ADMIN-PAYMENT-SECRET
      in: header